backend/requirements.txt
fastapi
uvicorn[standard]
clickhouse-connect
pydantic
python-dotenv
pandas
